code/python/satellite_image_project/requirements.txt
requests>=2.25.0
Pillow>=9.0.0
geopy>=2.3.0 